feat: add cancelled mail
