+++ apps/server/__tests__/__mocks__/Cancelled.ts
@@ -1,0 +1,1 @@
+export default 'Cancelled';
